Implement single fragment
